Move files to `src`
It is useful to have source files inside `src` for future organization
(e.g., tests, dockerization)
