+++ package.json
@@ -8,7 +8,7 @@
     "npm-audit-helper": "^3.1.0"
   },
   "scripts": {
-    "start": "node app.js",
+    "start": "node src/app.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "audit": "npm audit --audit-level=high --json | npm-audit-helper",
     "lint": "eslint src --max-warnings 0",
