+++ src/routes/index.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const router = express.Router();
+
+router.get("/", function(req, res) {
+    res.send("Hello world");
+});
+
+module.exports = router;
